// src/common/enums/regiones.enum.ts
export enum Regiones{
    Aguasturbias = "Aguasturbias",
    CiudadDeBanle = "Ciudad de Bandle",
    Demacia = "Demacia",
    ElFreljord = "El Fréljord",
    ElVacio = "El Vacío",
    IslasDeLasSombra = "Islas de la Sombra",
    Ixtal = "Ixtal",
    Jonia = "Jonia",
    Noxus = "Noxus",
    Piltover = "Piltóver",
    Shurimia = "Shurima",
    Targon = "Targón",
    Zaun = "Zaun"
}